\name{unknown.Np2}
\alias{unknown.Np2}
\docType{data}
\title{A Sample From an Unknown Bivariate Distribution 
}
\description{A 2500 x 2 matrix generated from 5000 normal(0,1) variables
}
\usage{data("unknown.Np2")}
\format{
  2500 x 2 matrix
}
\source{Generated by the author
}
\examples{
data("unknown.Np2")
}
\keyword{datasets}


