\name{unknown.Np4}
\alias{unknown.Np4}
\docType{data}
\title{A Sample From an Unknown Four-variate Distribution
}
\description{A 6000 x 4 x 1 array generated from 24000 normal(0,1) variables
}
\usage{data("unknown.Np4")}
\format{
  6000 x 4 x 1 array
}
\source{Generated by the author
}
\examples{
data("unknown.Np4")
}
\keyword{datasets}
