\name{unknown.Bp2}
\alias{unknown.Bp2}
\docType{data}
\title{A Sample From an Unknown Bivariate Distribution 
}
\description{A 3600 x 2 x 1 array generated from 7200 modified Bernoulli(0,1) variables. 
}
\usage{data("unknown.Bp2")}
\format{
  3600 x 2 x 1 array
}
\source{Generated by the author
}
\examples{
data("unknown.Bp2")
}
\keyword{datasets}
