\name{unknown.Bp4}
\alias{unknown.Bp4}
\docType{data}
\title{A Sample From an Unknown Four-variate Distribution 
}
\description{A 6000 x 4 matrix generated from 24,000 Bernoulli(0,1) variables
}
\usage{data("unknown.Bp4")}
\format{
  6000 x 4 x 1 array
}
\source{Generated by the author
}
\examples{
data("unknown.Bp4")
}
\keyword{datasets}
